+++ .gitignore
@@ -3,4 +3,4 @@ node_modules/
 index.js
 blockchain.js
 chainlog.js
-usercli.js
+login.js
