+++ README.md
@@ -2,7 +2,12 @@
 
 A simple blockchain and cryptocurrency wallet implemented in Node.js and TypeScript.
 
-# Usage
+## Development
+
+![Alt](https://repobeats.axiom.co/api/embed/18754e34717f50ca492537dfc8a38d49c0aacf1c.svg "Repobeats analytics image")
+
+<details>
+  <summary><h2>⚙️ Running locally for development</h2></summary>
 
 Start cloning Blockchain:
 ```
@@ -24,3 +29,5 @@ To start, run:
 ```
 npm start
 ```
+
+</details>
