--- chainlog.ts
@@ -1,3 +1,0 @@
-import { Chain } from "./blockchain";
-
-console.log(Chain.instance);
