+++ package.json
@@ -2,7 +2,7 @@
   "name": "blockchain",
   "version": "1.0.0",
   "description": "",
-  "main": "index.js",
+  "main": "./src/index.js",
   "scripts": {
     "test": "echo \"Error: no te,st specified\" && exit 1",
     "dev": "tsc -w",
